# server/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Privilege Game API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Player",
      "description": "Endpoints"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read_pets": "read your pets",
        "write_pets": "modify pets in your account"
      }
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/player/{id}": {
      "get": {
        "tags": [
          "Player"
        ],
        "description": "Gets an existing player.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Player's id."
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": [
          "Player"
        ],
        "description": "Updates an existing player by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "score",
            "in": "body",
            "description": "Player's score.",
            "required": true,
            "type": "Player",
            "schema": {
              "$ref": "#/definitions/updateScore"
            }
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": [
          "Player"
        ],
        "description": "Deletes an existing player by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/player": {
      "post": {
        "tags": [
          "Player"
        ],
        "description": "Creates a new player.",
        "parameters": [
          {
            "name": "newPlayer",
            "in": "body",
            "description": "Player's info.",
            "required": true,
            "type": "Player",
            "schema": {
              "$ref": "#/definitions/addPlayer"
            }
          }
        ],
        "responses": {}
      }
    },
    "/race/{id}": {
      "get": {
        "tags": [
          "Race"
        ],
        "description": "Get an existing race.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": [
          "Race"
        ],
        "description": "Updates an existing race by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "race",
            "in": "body",
            "description": "Race's info.",
            "required": true,
            "type": "Race",
            "schema": {
              "$ref": "#/definitions/updateRace"
            }
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": [
          "Race"
        ],
        "description": "Deletes an existing race by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/race": {
      "post": {
        "tags": [
          "Race"
        ],
        "description": "Creates a new race.",
        "parameters": [
          {
            "name": "newRace",
            "in": "body",
            "description": "Race's info.",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/addRace"
            }
          }
        ],
        "responses": {}
      }
    },
    "/level/{id}": {
      "get": {
        "tags": [
          "Level"
        ],
        "description": "Gets an existing level.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Level's id."
          }
        ],
        "responses": {}
      }
    },
    "/level": {
      "post": {
        "tags": [
          "Level"
        ],
        "description": "Creates a new player.",
        "parameters": [
          {
            "name": "newLevel",
            "in": "body",
            "description": "Level's info.",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/addLevel"
            }
          }
        ],
        "responses": {}
      }
    },
    "/question/{id}": {
      "get": {
        "tags": [
          "Question"
        ],
        "description": "Gets an existing question.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Question's id."
          }
        ],
        "responses": {}
      }
    },
    "/question": {
      "post": {
        "tags": [
          "Question"
        ],
        "description": "Creates a new question.",
        "parameters": [
          {
            "name": "newQuestion",
            "in": "body",
            "description": "Question's info.",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/addQuestion"
            }
          }
        ],
        "responses": {}
      }
    }
  },
  "definitions": {
    "player": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "jhon.doe@email.com"
        },
        "occupation": {
          "type": "string",
          "example": "analyst"
        },
        "is_leader": {
          "type": "boolean",
          "example": false
        },
        "skin_tone": {
          "type": "number",
          "example": 5
        },
        "gender": {
          "type": "number",
          "example": 3
        },
        "score": {
          "type": "number",
          "example": 0
        }
      },
      "required": [
        "email",
        "occupation",
        "is_leader",
        "skin_tone",
        "gender"
      ]
    },
    "addPlayer": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "jhon.doe@email.com"
        },
        "occupation": {
          "type": "string",
          "example": "analyst"
        },
        "is_leader": {
          "type": "boolean",
          "example": false
        },
        "skin_tone": {
          "type": "number",
          "example": 5
        },
        "gender": {
          "type": "number",
          "example": 3
        }
      },
      "required": [
        "email",
        "occupation",
        "is_leader",
        "skin_tone",
        "gender"
      ]
    },
    "updateScore": {
      "type": "object",
      "properties": {
        "score": {
          "type": "number",
          "example": 30
        }
      },
      "required": [
        "score"
      ]
    },
    "race": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "6183f7afbf78b91c1a0e2f9e"
        },
        "name": {
          "type": "string",
          "example": "privilege game"
        },
        "description": {
          "type": "string",
          "example": "privilege game description"
        },
        "total_players": {
          "type": "number",
          "example": 100
        },
        "has_prize": {
          "type": "boolean",
          "example": true
        },
        "prize": {
          "type": "string",
          "example": "A cash prize of U$100.000,00"
        },
        "lane_size": {
          "type": "number",
          "example": 100
        }
      },
      "required": [
        "id",
        "name",
        "total_players",
        "has_prize",
        "lane_size"
      ]
    },
    "addRace": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "privilege game"
        },
        "description": {
          "type": "string",
          "example": "privilege game description"
        },
        "total_players": {
          "type": "number",
          "example": 100
        },
        "has_prize": {
          "type": "boolean",
          "example": true
        },
        "prize": {
          "type": "string",
          "example": "A cash prize of U$100.000,00"
        },
        "lane_size": {
          "type": "number",
          "example": 100
        }
      },
      "required": [
        "name",
        "total_players",
        "has_prize",
        "lane_size"
      ]
    },
    "updateRace": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "privilege game"
        },
        "description": {
          "type": "string",
          "example": "privilege game description"
        },
        "total_players": {
          "type": "number",
          "example": 100
        },
        "has_prize": {
          "type": "boolean",
          "example": true
        },
        "prize": {
          "type": "string",
          "example": "A cash prize of U$100.000,00"
        },
        "lane_size": {
          "type": "number",
          "example": 100
        }
      }
    },
    "level": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 3
        },
        "boundary": {
          "type": "number",
          "example": 25
        },
        "name": {
          "type": "string",
          "example": "Intern"
        },
        "race_id": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "id",
        "boundary",
        "name",
        "race_id"
      ]
    },
    "question": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 15
        },
        "content": {
          "type": "string",
          "example": "Are your both parents graduated?"
        },
        "score": {
          "type": "number",
          "example": 5
        },
        "type": {
          "type": "number",
          "example": 1
        },
        "race_id": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "id",
        "content",
        "score",
        "type",
        "race_id"
      ]
    },
    "addQuestion": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string",
          "example": "Are your both parents graduated?"
        },
        "score": {
          "type": "number",
          "example": 5
        },
        "type": {
          "type": "number",
          "example": 1
        },
        "race_id": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "content",
        "score",
        "type",
        "race_id"
      ]
    }
  }
}